+++ wagtailmedia/models.py
@@ -5,12 +5,11 @@ import os.path
 
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
+from django.core.validators import MinValueValidator
 from django.db import models
-from django.db.models.signals import pre_delete
 from django.dispatch import Signal
-from django.dispatch.dispatcher import receiver
 from django.urls import reverse
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from taggit.managers import TaggableManager
 from wagtail import VERSION as WAGTAIL_VERSION
@@ -48,7 +47,13 @@ class AbstractMedia(CollectionMember, index.Indexed, models.Model):
     file = models.FileField(upload_to='media', verbose_name=_('file'))
 
     type = models.CharField(choices=MEDIA_TYPES, max_length=255, blank=False, null=False)
-    duration = models.PositiveIntegerField(verbose_name=_('duration'), help_text=_('Duration in seconds'))
+    duration = models.FloatField(
+        blank=True,
+        default=0,
+        validators=[MinValueValidator(0)],
+        verbose_name=_('duration'),
+        help_text=_('Duration in seconds'),
+    )
     width = models.PositiveIntegerField(null=True, blank=True, verbose_name=_('width'))
     height = models.PositiveIntegerField(null=True, blank=True, verbose_name=_('height'))
     thumbnail = models.FileField(upload_to='media_thumbnails', blank=True, verbose_name=_('thumbnail'))
@@ -113,6 +118,11 @@ class AbstractMedia(CollectionMember, index.Indexed, models.Model):
         from wagtailmedia.permissions import permission_policy
         return permission_policy.user_has_permission_for_instance(user, 'change', self)
 
+    def clean(self, *args, **kwargs):
+        super().clean(*args, **kwargs)
+        if not self.duration:
+            self.duration = 0
+
     class Meta:
         abstract = True
         verbose_name = _('media')
@@ -149,14 +159,6 @@ def get_media_model():
             settings.WAGTAILMEDIA_MEDIA_MODEL
         )
     return media_model
-
-
-# Receive the pre_delete signal and delete the file associated with the model instance.
-@receiver(pre_delete, sender=Media)
-def media_delete(sender, instance, **kwargs):
-    # Pass false so FileField doesn't save the model.
-    instance.file.delete(False)
-    instance.thumbnail.delete(False)
 
 
 media_served = Signal(providing_args=['request'])